Implement Discount class and changed names and types of variables in OfferItem class
